Split the desktop main process into scripts modules.
